In Web.JS, rename ILogger to Logger to match TypeScript conventions
